requirements.txt
pytest==8.2.2
pytest-xdist==3.6.1
pytest-html==4.1.1
allure-pytest==2.13.5
pyyaml==6.0.1
playwright==1.45.0
pytest-playwright==0.5.1
pandas==2.2.2